Scraper/requirements.txt
beautifulsoup4
pyodbc
python-dotenv
scrapy